Move helper into different file/dir
